Conflicts:
	src/Trades/CoinExchange.Trades.Port.Adapter.Rest/Resources/OrderController.cs
	src/Trades/CoinExchange.Trades.Port.Adapter.Rest/Resources/TradeController.cs

+++ src/Trades/CoinExchange.Trades.Port.Adapter.Rest/Resources/OrderController.cs
@@ -133,22 +133,20 @@ namespace CoinExchange.Trades.Port.Adapter.Rest.Resources
                     string[] auth = headerParams.ToList()[0].Split(',');
                     apikey = auth[0];
                 }
-                // ToDo: In the next sprint related to business logic behind RESTful calls, need to split the ledgersIds comma
-                // separated list
                 object value = _orderQueryService.GetOpenOrders(new TraderId(int.Parse(Constants.GetTraderId(apikey))),
                     queryOpenOrdersParams.IncludeTrades, queryOpenOrdersParams.UserRefId);
 
                 if (value is List<OrderRepresentation>)
                 {
-                    List<OrderRepresentation> openOrderRepresentation = (List<OrderRepresentation>)value;
+                    List<OrderRepresentation> openOrderRepresentation = (List<OrderRepresentation>) value;
                     return Ok<List<OrderRepresentation>>(openOrderRepresentation);
                 }
                 else if (value is List<OrderReadModel>)
                 {
-                    List<OrderReadModel> openOrderList = (List<OrderReadModel>)value;
+                    List<OrderReadModel> openOrderList = (List<OrderReadModel>) value;
                     return Ok<List<OrderReadModel>>(openOrderList);
                 }
-
+                
                 return NotFound();
             }
             catch (Exception ex)
@@ -180,8 +178,7 @@ namespace CoinExchange.Trades.Port.Adapter.Rest.Resources
                     string[] auth = headerParams.ToList()[0].Split(',');
                     apikey = auth[0];
                 }
-                // ToDo: Get the Trader ID after authentication using API key.
-                object orders = _orderQueryService.GetClosedOrders(new TraderId(Int16.Parse(Constants.GetTraderId(apikey))),
+                object orders = _orderQueryService.GetClosedOrders(new TraderId(int.Parse(Constants.GetTraderId(apikey))),
                     closedOrdersParams.IncludeTrades, closedOrdersParams.UserRefId, closedOrdersParams.StartTime,
                     closedOrdersParams.EndTime, closedOrdersParams.Offset, closedOrdersParams.CloseTime);
